Add files via upload
This program uses accesses a previously saved to do list from the files. It stores each task with the name of the task(what it entails), the importance of it and the date it has to be done by. You can mark items as completed and add new things to the list, then copy and paste it to the file, so the to do list is not lost at the end of the program.